requirements.txt
Django==2.1.7
psycopg2
django-multitenant
Pillow
django-debug-toolbar
djangorestframework
requests
